changed colour of histogram

+++ classy_text/vectorize_data.py
@@ -51,7 +51,7 @@ def ngram_vectorize(train_texts, train_labels, val_texts):
     # Create keyword arguments to pass to the 'tf-idf' vectorizer.
     kwargs = {
             'ngram_range': NGRAM_RANGE,  # Use 1-grams + 2-grams.
-            'dtype': 'int32',
+            'dtype': 'float64',
             'strip_accents': 'unicode',
             'decode_error': 'replace',
             'analyzer': TOKEN_MODE,  # Split text into word tokens.
